moved AnswerState enum into its own file
