+++ enums/AnswerState.ts
@@ -1,0 +1,5 @@
+export const enum AnswerState {
+    Correct = 'CORRECT',
+    Incorrect = 'INCORRECT',
+    Unanswered = 'UNANSWERED'
+}
